backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
colorthief==0.2.1
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 